(#7015) | refactor lazy_repeat_array
