+++ docs/source/api_reference/syft.core.node.common.rst
@@ -37,6 +37,7 @@ syft.core.node.common
    syft.core.node.common.client
    syft.core.node.common.client_manager
    syft.core.node.common.exceptions
+   syft.core.common.lazy_repeat_array
    syft.core.node.common.metadata
    syft.core.node.common.node
    syft.core.node.common.node_manager
