+++ packages/syft/tests/syft/core/tensor/adp/gamma_tensor_test.py
@@ -9,7 +9,7 @@ import syft as sy
 from syft.core.adp.data_subject_list import DataSubject
 from syft.core.tensor.autodp.gamma_tensor import GammaTensor
 from syft.core.tensor.autodp.phi_tensor import PhiTensor as PT
-from syft.core.tensor.lazy_repeat_array import lazyrepeatarray as lra
+from syft.core.common.lazy_repeat_array import lazyrepeatarray as lra
 
 
 @pytest.fixture
