+++ packages/syft/tests/syft/core/tensor/private_test.py
@@ -6,7 +6,7 @@ import pytest
 import syft as sy
 from syft.core.adp.data_subject import DataSubject
 from syft.core.tensor.autodp.phi_tensor import PhiTensor as PT
-from syft.core.tensor.lazy_repeat_array import lazyrepeatarray as lra
+from syft.core.common.lazy_repeat_array import lazyrepeatarray as lra
 from syft.core.tensor.tensor import Tensor
